Move EngineConfig to a Static Global
EngineConfig is read-only and logically makes sense as a static as it
last the life of the program, so this moves it to a global static that
can be accessed from anywhere in the program instead of just in the Bevy
world.

This is also going to be required for scripting support so that the
asset directory will be readable from asset loaders without borrowing the
Bevy world.